Huge rewrite of Hex and stencil classes into general CellSpace class
Just like with OpenFOAM's standard VectorSpace base class of vectors and
tensors, cell-related primitives are now based on a similar CellSpace class. The
CellSpace class is an array wrapper around OpenFOAM's standard primitives
(label, scalar, vector, etc), and transmutes all operations and functions as an
iteration over the array. The benefit is that all functions and operations are
defined only once (in CellSpaceI.H), apart from class-specific ones. The
CellSpace class is different from the VectorSpace class in that, in principle,
no reduction of the CellSpace type itself occurs as a result of an operation,
but only the contained type is reduced. For example, the inner product between
two CellSpace types each templated for an OpenFOAM primitive is again a
CellSpace type templated for the inner product of the two primitives. I.e.,
CellSpace<vector, ...> & CellSpace<vector, ...> := CellSpace<scalar, ...>.

In recode process, Hex was renamed to FaceSpace<Type>, so that the previously
called hexScalar class has become faceScalar [ = FaceSpace<scalar> ] and
hexVector has become faceVector. This naming seems more intuitive.

We can now easily extend the stencil class to, for example, a wideStencil class,
which would contain coefficients belonging to a larger number of neighbors.
Also, we can now easily extend the FaceSpace class to, for example an EdgeSpace
class, to form the edgeScalar and edgeVector types. This is not yet done.