Added method to automatically align bricks in structured meshes
When a mesh is structured, bricks can be aligned. This avoids unnecessary data
transformation. A method is now added that performs the brick alignment. All
bricks are aligned w.r.t. the first brick that is found in the brick topology
map while traversing the map first in i, then in j and finally in k. The
transformation tensor is found by tracing the shortest path between the brick to
be aligned and the first brick.