Mesh direction rewrite
A mesh direction contains a block with size = internal cells + ghost cells +
padding. In those cells, we can mark a sub-block as 'active cells', i.e., cells
for which an equation should be solved. The mesh direction now calculates the
start and end index of the active cells, which is then used in the forAllCells
macro to iterate over all active cells. This avoids redundancy in staggered
meshes, where two processors contain the same cells. It also simplifies boundary
start and end index calculation.

Also, the decision was made not to support staggered unstructured meshes. A
check is now added to the meshField constructor, to support staggered mesh
construction only on meshes with a structured brick topology.