Added Poisson solver class
The PoissonSolver class is a special solver dedicated to solving the equation
dx/dt = div(lambda*grad(x)) + b. The 'default' Poisson solver, which is selected
when the specified solver type is not found in the Poisson solver lookup table,
simply constructs the Poisson equation manually and uses a regular solver to
solve it. Using the Poisson solver class, we can implemented special solvers
dedicated to Poisson equations.