Rewrote VTK input/output
The IO class has been rewritten to use the legacy VTK format, so that there is
no external dependence on VTK or general XML libraries. Parallel output is now
combined into a single .vtk file. Time series data is collected in a .vtk.series
file, which can be loaded by ParaView as of version 5.5 (see
https://gitlab.kitware.com/paraview/paraview/blob/v5.5.0/Documentation/release/ParaView-5.5.0.md#json-based-new-meta-file-format-for-series-added).
Also an IO test application is added, which constructs all types of fields on a
box and pipe mesh, then writes them, clears them, reads them again and checks if
the read data is correct. This is done both in ascii mode and binary mode. The
ascii mode is substantially slower than the binary mode, so write mode is set to
binary in all cases by default.