Reviewed brick, brick topology map and processor map coordinate systems
A right-handed coordinate system is now enforced when constructing the brick.
This was already done implicitly by a negative volume check, which achieved the
same. But the check is now moved to the constructor. Also the alignment of the
brick topology map, and the processor map that follows from it, are reviewed. We
have the following methodology:

1) Bricks must be right-handed (i.e., dot(v1 x v2, v3) > 0 with v1, v2 and v3
the x, y and z vertex vectors from the brick's origin), but may have local
coordinate systems that are not aligned from brick to brick.

2) For unstructured brick topologies, there will always be a pair of two facing
bricks that are not aligned.

3) For structured brick topologies, bricks are automatically aligned so that on
a face shared by two bricks the local coordinates are the same. Alignment occurs
w.r.t. the first brick in the bricks list in briscolaMeshDict.

4) For structured brick topologies, we can build a brick map. The brick map
construction is initiated w.r.t. the first brick and adheres to the first
brick's coordinate system. Thus, the brick topology map is aligned with the
first brick's coordinate system and therefore also with all other bricks
(because the bricks on structured brick topologies are automatically aligned)

5) Note that the way the code does this is:

    a) Leave bricks as-is and construct brick topology map

    b) If a brick topology map can be constructed, it means that bricks are
    structured

    c) If bricks are structured, align all bricks and then recreate the topology
    map, which will now contain no transformation tensors other than the
    identity tensor

   So the topology map is actually created twice.

6) For structured brick topologies, we can also build a processor map. The
processor map is constructed by adhering to the local coordinate system of the
brick topology map, and thus also that of the first brick.

Conclusion is that on structured brick topologies, the processor map reflects
the organization of processors in the same coordinate system as that of the
organization of bricks and also that of the vertcies of all the bricks. Of
course, this local coordinate system is not necessarily aligned with the global
coordinate system. This depends on the choice made for the first brick. It's
probably most intuitive to align this one with the global coordinate system, but
it is not required (and not even possible for skewed bricks).