Rewrite of mesh with runtime selection
There are several types of meshes. From general to most constrained, it can be
unstructured, structured, rectilinear or uniform. A uniform mesh can be regarded
as rectilinear, structured and unstructured. A rectilinear mesh as structured
and unstructured, etc. Each mesh type has its own specific methods. For example,
a uniform mesh is defined by three cell sizes in each direction, a rectilinear
mesh by three cell size arrays in each direction, etc. Therefore, it is natural
to define classes for each mesh type. Because we only know what kind of mesh
we're dealing with at runtime, the mesh class needs to have a 'virtual
constructor' using the factory method.

Drawback is that fvMesh can no longer be derived from the mesh class but must
have a pointer to a virtual mesh object, not knowing its concrete type. Thus,
some member functions of mesh must be put explicitly in fvMesh.

With this rewrite, we can 'upcast' a mesh object to a derived type like
uniformMesh. The mesh.cast<meshType>() member function takes care of this, and
checks if the upcast can be done (i.e., an unstructured mesh cannot be upcast to
a uniform mesh). If it fails, it gives an error. This is useful for methods
(like the FFT solver) which require the mesh to be of a certain type. If the
upcast succeeds, a new set of member functions becomes available belonging to
the specific mesh type.