Revision of internal/active cell calculation
Internal cells are now defined as all the cells that belong to a processor. On
shifted meshes, the cells that are shifted on the boundary are shared across a
master and slave processor, and are considered as internal cell on the master
processor but as ghost cell on the slave. This is defined by the part patch, and
is therefore independent of the field.

For Dirichlet boundary conditions, which is applied field-specific, internal
cells that are on the boundary due to mesh shift are 'slave', i.e., they do not
have to be computed as part of the linear system. Such cells are internal, but
inactive.

The forAllCells macro iterates over all internal cells. The forAllActiveCells
macro iterates over all active cells, which thus excludes Dirichlet boundary
cells that are on the boundary.