Added part patch calculation for internal boundaries on brick links
On a brick face links we can have edge or vertex links on internal processor
boundaries. On brick edge links we can have vertex links on internal processor
boundaries. These are now added.

Also a function addPartPatch was added which creates a dictionary and adds to
the list of part patches.