Added cell search function
This uses a tree search, by first looking at the coarsest cells. The
insideCell() algorithm is very generic and works for any type of hexahedron.
This makes it quite expensive, but since we're using a tree search the function
isn't called too often.

Also, mesh specializations (uniform/rectlinear) can overload the search
algorithm with faster algorithms (to be implemented).