Removed 'active cell' methodology
Instead, boundary conditions that constrain values on boundaries (Dirichlet,
empty) do so by modification of the stencil in the
boundaryCondition::correctSystem function. This function was already being
called anyway, for eliminated boundary values.

This change removes the need to adjust the mesh direction internal cell start
and end indices, which simplifies the code. The forAllCells iterator is now
independent of the boundary condition and, therewith, independent of the field
too.