Added deep/shallow field functionality
In the constructor of a meshField, it can be specified that the field is shallow
(not deep), meaning that only the top level will be created. This will save some
memory (about 14%) and some CPU time.

By default, fields are currently deep. It is intended to switch to shallow
fields by default, but this will require some more work.