Added direct access operators to field/level
This simplifies indexing to cells for fields that only have one level or levels
that only have one directions. In most cases the user is not really concerned
with levels and for colocated fields directions are irrelevant. So, we now have:

    field(l,d,i,j,k) -> field[l][d](i,j,k)
    field(d,i,j,k) -> field[0][d](i,j,k)
    field(i,j,k) -> field[0][0](i,j,k)

    level(d,i,j,k) -> level[d](i,j,k)
    level(i,j,k) -> level[0](i,j,k)

Also, iterator macros have been added to reduce the number for-loops that have
to be programmed. For example:

    forAllLevels(field, l, d, i, j, k)
    {
        field(l,d,i,j,k) = ...
    }

    forAllDirections(field, d, i, j, k)
    {
        field(d,i,j,k) = ...
    }

    forAllCells(field, i, j, k)
    {
        field(i,j,k) = ...
    }

    forAllDirections(level, d, i, j, k)
    {
        level(d,i,j,k) = ...
    }

    forAllCells(level, i, j, k)
    {
        level(i,j,k) = ...
    }

The simplification of indexing is tested in Test-meshField and Test-meshLevel
but is yet to be implemented in the rest of the code.