Reorgaized vof schemes
The original vof class is now a base class with pure virtual functions. The
implementation of the advection is now in the splitAdvection class. This will
allow for other advection methods in the future, e.g., an unsplit one or even an
algebraic one.

The vof scheme type must now be selected in the dictionary, so the Vof cases
were also adjusted accordingly.