Further reorganization of Vof code and added twoPhase library
The splitAdvection class is now derived from the geometricVof class which is
derived from the vof class in turn. The vof test app was renamed to
geometricVof. A twoPhase library was added which defines two-phase mixture
properties. A basic 'incompressibleVof' two-phase model was added, which
currently computes rho and mu from a volume fraction and harmonic volume
fraction weighting, respectively.