Added cell data exchange method
Using the cellDataExchange class, we can exchange cell data across neighboring
processors. The cellDataExchange object must be constructed using a list of cell
indices that are not in the local processor part. For example, on a part that
has 32^3 cells, the index (33,2,2) would be in the index (2,2,2) in the right
neighbor. Once the object is constructed, it can be called on a field to
exchange data of that field. See the cellDataExchange and
cellDataExchangeUnstructured test applications.