Fix in interpolationWeights
It seems that rounding the interpolation weight to an integer multiple of the
set tolerance helps to consistently determine in which cell a point belongs. The
difficulty is that points sitting very close to a face may not be recognized to
be within either cell on each side of the face. Rounding seems to help with this.