Schemes cleanup and performance improvements
Schemes are restructured a bit and are extended to take more arguments. For
example, the Laplacian scheme now also takes a constant coefficient as argument,
which avoids the construction of a whole face field for the viscosity. Also, the
implicit Laplacian, divergence and source schemes now have methods where they
directly add to a linear system, instead of creating a new one and returning
that as a tmp. This avoids allocating memory and copying that memory, giving a
small performance improvement.