Further fixes in two-phase solvers
The way that v = 1/rho is calculated at the face is important. For the colocated
two-phase solver this needs to be an interpolation of the cell-centered v. For
the staggered solver, the colocated face value of v needs to be consistent with
the staggered cell-centered value. Since staggered cell centers coincide with
colocated faces, we can therefore directly copy values from the staggered field
to the colocated faces.

It seems that the restriction scheme of v at the face is best set to a face
average. For the cell-centered field v, the restriction should be a harmonic
average. This is the same for both staggered and colocated solvers.

Also a reduced pressure formulation was introduced, but this is still
experimental and needs to be tested. It can be used by setting 'reduced' to true
in the briscolaTwoPhaseDict case file.