Added split Poisson equation solver
Based on the work of Dodd & Ferrante, an approximate constant-coefficient
Poisson equation is generated and then solved with another Poisson solver. The
implementation is fully contained within the splitPoissonSolver class. The
PoissonSolver class now has the option to compute the flux, so that it is
consistent with the way that the pressure equation is solved.