Pressure gradient is stored in solvers
This way it can be reused in the next time step, without having to reconstructor
it. The pressure gradient is needed for an implicit momentum equation system,
which we currently have due to the Crank-Nicolson treatment of the viscous term.
In the future, when using RK schemes, this may no longer be needed when using
fully explicit RK schemes.