Further improvements to the two-phase solvers
The colocated two-phase solver is improved by handling the surface tension flux
in a better way. It is now part of the pressure gradient G.

The staggered two-phase solver no longer integrates the surface tension term
using Adams-Bashforth, but explicitly at t^{n+1}. Since the surface tension term
depends fully on the alpha prediction, and since the alpha prediction is
second-order, so is the surface tension term.

Both the colocated and staggered solvers give very similar results for the
Hysing case.