Improved automatic boundary condition correction
When schemes with spatial operations are given a meshField to operate on, they
require the BCs to be updated. In principle, stored fields (like p and U) always
have their BCs corrected. Other fields, like temporary ones, don't automatically
correct the BCs. Thus, the spatial operation requires a BC correction on tmp
fields. This is now implemented.

An example of where this is important is:

    stagT(ex::grad(U))

The ex::grad(U) operation returns a field without BC correction. The stagT
function (staggered transpose) performs interpolation and thus needs the BCs to
be corrected. The stagT(tmp<...>) function performs the boundary condition
correction now automatically. Another example is

    ex::div(mu*ex::faceFlux(T(ex::grad(U)))).