Added Uc calculation to staggered solvers again
After some thought, I think it's best to have Uc being calculated at the end of
each time step. Not only is it more convenient for visualization, I also expect
that Uc will be needed by other models or FOs. It is already the case for the
budget primitives.

The read mode of Uc is now READ_IF_PRESENT. This means that if BCs are specified
for Uc, they are used. This is important of Uc is operated on with spatial
schemes, for example in the budget primitives FO.