correctBoundaryConditions() now also corrects IBCs on the first level
A call to correctBoundaryConditions() now also updated the immersed boundary
conditions on the first (finest) mesh level. This guarantees that the immersed
boundary condition correction is always done before the normal boundary
condition correction. This is important for the consistent transfer of forced
immersed boundary cell to neighbors.

If the immersed boundary conditions are also to be corrected on coarser levels,
this must be done manually by calling correctImmersedBoundaryConditions().

A drawback of this approach is that every fields needs to have the IBCs set.
This means that if no boundaryConditions dictionary is found in a field, empty
immersed boundary conditions must be initialized. Fortunately, this is tiny
overhead.