Small fix in interpolations
Once we've determined the cell that a point is in, there's no need to call the
interpolationWeights function with the insideOnly parameter set to true. In some
cases the point may still fall just outside the cell with a distance in the
order of machine precision, but that's not really a big problem. The weights
should be allowed to be slightly negative in that case.

This fix resolves issues with interpolating to points that sit exactly (i.e.,
within machine precision) on a cell face.