Added dedicated Crank-Nicolson/Adams-Bashforth (CNAB) solvers
Two special solvers (briscolaColocatedCNAB and briscolaStaggeredCNAB) are added
that implement the CNAB IMEX scheme, see the paper of:

    Ascher, U. M., Ruuth, S. J., & Wetton, B. T. (1995). Implicit-explicit
    methods for time-dependent partial differential equations. SIAM Journal on
    Numerical Analysis, 32(3), 797-823.

By default, the Modified CNAB (MCNAB) algorithm is used, but it also possible to
revert to the classic CNAB by setting the modified switch to false in the solver
dictionary. The MCNAB has slightly superior stability and accuracy properties.