Big update on direct solvers
Added a linearSystemAggregation class that aggregates the linear system over a
number of parts. The direct solver can then be called for the aggregated part,
and one needs to iterate on the boundary conditions. However, this is likely to
be much faster than aggregating the full matrix on one processor, because:

1) Direct solvers have poor scaling, >= O(N^2) for sparse linear systems. So
   splitting up the system will have superlinear scaling.

2) The matrix factorization only needs to occur once. After that, each boundary
   update only requires the solution to an LU system which is very fast.

Also the iterative BiCGSTAB is added as Eigen solver. This is not a direct
solver, but since we have to iterate on the aggregated boundaries anyway, this
is ok.