Added VSCode settings to enable C/C++ IntelliSense add-on
The project is configured such that the root of the project is *not* included,
because this may slow down the C/C++ add-on when there's a lot of case data.
Also error squiggles are disabled because none of the dependencies are included
in the include path, which would give a lot of errors otherwise.

By default the following OpenFOAM directories are included:

"/software/OpenFOAM/OpenFOAM-9/src/OpenFOAM/**",
"/software/OpenFOAM/OpenFOAM-9/src/Pstream/mpi/**",
"/software/OpenFOAM/OpenFOAM-9/src/OSspecific/POSIX/**",
"${FOAM_SRC}/OpenFOAM/**",
"${FOAM_SRC}/Pstream/mpi/**",
"${FOAM_SRC}/OSspecific/POSIX/**",

The first three are hard-coded paths that work on Vinci. The second three should
work when OpenFOAM is loaded via .bashrc or .bash_profile etc.