Removed LowerFaceSpace and symmStencil
The asymmetric LowerFaceSpace and symmStencil primitive types and derived
objects were removed. While these primitives exploit symmetry and thus cut
memory storage in half, their use has quite an impact on the code while not
being significantly faster (this was tested). For example, using LowerFaceSpace
the calculation of the divergence (i.e. the continuity error) used to be:

```
     forAllCells(contError, i, j, k)
        for (int fd = 0; fd < 3; fd++)
            contError(i,j,k) +=
                phi(i,j,k)[fd]
              - phi(upperNei(i,j,k,fd))[fd];
```

while now it is

```
     forAllCells(contError, i, j, k)
        for (int f = 0; f < 6; f++)
            contError(i,j,k) += phi(i,j,k)[f];
```

So, this avoids the calculation of the upper neighbor index in direction fd and
also improves memory access alignment. The cost is that each flux is stored
twice, for both cells adjacent to the face. Using the forAllFaces macro, each
flux only needs to be computed once and stored twice.

Another benefit is that everything is stored in internal cells, so that handling
ghost cells properly is not so important anymore. In fact, it was tested to have
all meshDirection field functions operate on internal cells only and not on the
underlying block data which includes ghosts. This now works without error, but
is not cheaper because the direct operation on the block data, while larger in
size, avoids the need for index calculation and is perfectly aligned with
memory.

The removal of symmStencil mainly makes implementation more straight-forward.
For example, before it was needed to get the full matrix row stencil first
before building the matrix needed for direct solvers. This step can now be
avoided because we only use the full stencil type.