Bugfix in cellDataExchange
It turned out not to be sufficient to identfy the neighboring processor in
cellDataExchange by just the processor number, because a situation may occur in
which two faces may be neighboring the same processor, through a periodic and
parallel boundary. So to find the correct neighbor, the boundary offset must be
communicated too.