Added reduced pressure formulation
For periodic simulations, a reduced pressure formulation will be needed. This is
now implemented, and can be switched on by the corresponding parameter in the
briscolaTwoPhaseDict dictionary.

Also the TwoPhaseModel class has now a flux() member function which sums and
returns the buoyancy and surface tension interfacial fluxes. This simplifies
things in the two-phase solvers.