Added allStages and lastStage pressure modes to RK scheme
The pressure mode (pMode) can be set in the RK scheme properties in the
briscolaSchemeDict file, like this:

```
rkScheme
{
    scheme  <RK scheme>;
    pMode   <lastStage or allStages>;
}
```

By default, the pMode is set to allStages. When it is manually set to lastStage,
only on the last stage a pressure equation is solved and the velocity is
corrected with a pressure gradient. This can be seen as a fractional step:

```
du/dt = F + P ->
    u*      = u^n + dt*F,
    u^(n+1) = u*  + dt*P,
```

with F the convection/diffusion part and P the pressure part.

Also two special RK schemes were added: RK3CN and RK4CN. These are schemes that
apply RK3/RK4 for the convective term without diffusion, and only in the final
stage they solve for diffusion with the Crank-Nicholson scheme. When pMode is
set to lastStage, this reduced to a tri-fractional step method:

```
du/dt = C + D + P ->
    u*      = u^n + dt*RK3(C),
    u**     = u*  + dt*CN(D),
    u^(n+1) = u** + dt*P,
```

with convection C and diffusion D. The function RK3 applies RK3 and the function
CN applies Crank-Nicholson.