Improved buoyancy formulation
It seems that the best way to get rid of the static pressure gradient in
periodic domains is by subtracting rhoMean*g from the right-hand side of the
momentum equation. The static pressure jump can be shown to be

    dp = rhoMean*g*H

Thus, the corresponding gradient is dp/H = rhoMean*g. For reference, this is
what Tryggvason et al. write about it in the book 'DNS of Gas-Liquid Multiphase
Flows':

"For theoretical work, periodic boundaries are often very attractive, since it
is not necessary to deal with the often difficult task of specifying appropriate
in- and out-flow boundary conditions. The only new consideration is that if the
domain is periodic in the direction of gravity, then it is necessary to add a
body force equal to rho_av*g, where rho_av is the average density of the whole
domain, to prevent a uniform acceleration of the fluid. This force gives rise to
a hydrostatic pressure gradient, but since it results in the conservation of
momentum in the domain - not the average velocity - it does not correspond
exactly to a container with a rigid bottom."