Updated the way that brick vertices are defined in briscolaMeshDict
To be more consistent with the way that faces are defined, brick vertices are
now to be provided as a list of vertex indices, thus getting rid of all the
brackets. It's a bit more consistent and cleaner.

A regular expression was used to replace all instances with the new formatting.