Added globalUniform check to rectilinearMesh, needed for FFT solver
As it stood, the FFT solver only checked if each processor's partial mesh was uniform in at least two directions. However, the case where multiple uniform partial meshes form a non-uniform global mesh was not being checked for. This check was now added to the FFT solver to make sure that an error is shown when the FFT solver is used for any global mesh with at least two non-uniform directions.

Also added brief description to FFT solver header.