Added geometric objects as primitives
Added the tetra, hexa, piped and tessellation classes, which are geometrical
objects and can be used to consistently compute geometric properties. Mesh
metrics are now computed from these objects. Truncating a geometrical object is
now possible ussing the tessellation class, which is simply a list of
tetrahedrals. Decomposing any geometrical object into a tessellation allows it
to be truncated and allows for simple volume computation. This will be needed
for advanced geometrical advection algorithms.