Restructured two-phase library and added twoPhaseMultiVof
Surface tension and interface normal schemes are now moved from the twoPhaseModel level to the twoPhaseSolver level. In twoPhaseMultiVof, the solver has a PtrList of surface tension and normal schemes corresponding to the multiple marker fields. These are then summed to get one total surface tension flux field.

Currently, twoPhaseMultVof is hard-coded to two alpha fields, but this is to be extended to n fields with automatic creation of fields based on CCL and coalescence suppression.