Fixed memory leak in PETSc
PETScExternalSolver::solve created a temporary Vec but never destroyed it. This was the main leak as this accumulates at every solve call. Another leak was fixed in PCLUPETScSolver.C but this leak did not accumulate over time so was less problematic.

There are still some small leaks in PETSc like communicators not getting destroyed but again these only appear at finalization and are therefore (probably) harmless.