Reorganized Dirichlet, Neumann and empty BCs
It is more convenient for derived boundary conditions to define everything
within the same class, i.e., to not have stagDirichlet, stagNeumann or
stagEmpty. To realize this, we need to partially specialize the full Dirichlet,
Neumann and empty BC classes. To avoid lots of code duplication, base classes
were created that do most of the work. Then, then partial class specialization
of the colocated and staggered mesh types become very small, and only the
eliminateGhosts and evaluate functions need to be implemented for each mesh
type.