Added fake boundaries
For some boundary conditions we will need to be able to do global operations.
For example, if we need to compute a flow rate or for mapped boundary conditions
to perform interpolation. Global operations are only needed for boundaries on
patches that are not empty (so, not on parallel/periodic boundaries).

To enable