Reviewed PETSc memory usage
Whenever a PETSc create function is called, the object (Mat, KSP) is first
destroyed using PETSc's Destroy function. This should remove memory leaks,
because memory is only created once it is cleaned first.

Valgrind will still report errors on exit. Even if the remaining allocated PETSc
memory would be properly destroyed in the destructor of objects, valgrind errors
still prevail as per
https://lists.mcs.anl.gov/pipermail/petsc-users/2014-January/020329.html
pointing to the PetscInitialized() call and associated with MPI which is not
valgrind clean. Therefore it is pointless to attempt to get rid of finalization
errors.