Added faceDotGrad scheme
In two-phase solver the second part of the viscous term, i.e.,

div(mu*(grad(U))^T)

is not zero because mu changes across the interface. Thus, it needs to be
accounted for. The first part of the viscous term is

div(mu*grad(U))

and amounts to taking 'the spatial gradient dotted by the face normal' for each
component of U. This is what the faceGradient scheme does. Conversely, the
second part of the viscous term amounts to taking a spatial gradient of 'each
component of U dotted by the face normal'. For this, we need to have the
counterpart of the faceGradient scheme that does this. Such a scheme is now
implemented and called the a faceDotGradient scheme, for lack of a better name.

For the colocated scheme, taking a spatial gradient at a face requires one
normal direction gradient and two tangential direction gradients. In turn, this
requires interpolation which can be quite expensive.

For the staggered scheme the story is a bit simpler because in the tangential
directions the values of the dotted velocity components are readily available
from the other two staggered directions and conveniently positioned across the
face.

The faceDotGrad function is used now to compute the second part of the viscous
term in all four two-phase solvers (colocated/staggered, RK/CNAB). It is
implemented in the same way as the buoyancy source term, i.e., it is not handled
like the implicit Laplacian because it can only be computed explicitly. And
since it is associated with gradients in mu and thus in alpha, it probably makes
sense to handle it in the same way as buoyancy which is also associated with
gradients in alpha.