Centralized memory management of tmp objects
Tmps are now consistenly created using objectType::New functions implemented for
meshField/Level/Direction and linearSystem. This centralized the memory
management of objects. It allows for object caching, registering objects in a
special cache registry and therewith letting them survive destruction of tmps.
If a similar tmp is created, it is simply created from the cached object. This
potentially gives a performance gain, because we don't have to allocate memory
all the time. However, some testing showed that there is negligible gain when,
for example, caching large data chunks such as linear systems. It seems that
reusing the memory (probably loading it into processor cache) is just as
expensive as allocating new memory. So, the caching mechanism is not in use now,
but it is left in place for potential future use.

Along with these changes, also some operators of linearSystem have been
improved/added.