Simplified meshField constructor
No need to check if boundary condition file exists because this is already done
in IOdictionary of which meshField is derived.

Also a check in IO was added to see if all fields that have the read option
MUST_READ are actually present in the data files and read.

Finally, an if-statement was removed that prevents staggered fields on
unstructured meshes. We will also need staggered fields on unstructured meshes
soon.