Added FastPtrList class
This class is derived from PtrList and enherits all its features, but overrides
the operator[] member function to avoid a costly check of the pointer validity.
This makes the class much faster when iterating over objects contained by the
list. Rule of thumb is to use FastPtrList for objects of which it can be
expected that they are iterated, such as data fields or lists.