Added SoA storage for metrics
This is implemented to do some testing on the efficiency of SoA storage. It
seems to make face-based operations faster. This commit is a temporary one, as
the next step will be to make SoA storage the default for face fields and
probably for stencils too. The changes in the schemes will remain.