Added meshFieldSet class
This class is basically a templated wrapper around
FastPtrList<meshField<Type,MeshType>>, with the number of mesh fields defined by
a template parameter. It implements constructors and operators. Short-hand
functions like max, mag, etc., are yet to be defined (maybe they are not
necessary, tbd). The class is useful for storing for example face data or
stencil data, needed to improve performance.