Integration of faceField
This is a huge update replacing FaceSpace fields with faceFields with SoA
storage. This required all schemes that use face values to be rewritten. Also
boundary conditions and restriction schemes had to be rewritten. A new
faceRestrictionScheme class was introduced.

The default way of storing metrics is via SoA too. AoS metrics can still be
obtained, but this is on-demand to reduce memory usage.

The IBM part still requires a thorough review, rewriting things with forAllFaces
loops where possible.

Tests and validation cases pass, but there may still be bugs because much has
changed.