Automatic data alignment
Bricks are automatically rotated such that the highest number of cells
corresponds with the third local coordinate system's direction. This is
beneficial because it increases contiguous memory density. Particularly for
two-dimensional geometries this can make a signficiant difference.

To support automatic alignment, the user must be able to specify input values
independent of the local coordinate system. Instead, all inputs are now expected
to be based on the global coordinate system. This includes Dirichlet/Neuman
boundary condition values/gradients and the gravity vector. As a result, inputs
are no longer needed to be specified as a list of directional values. This
unifies colocated and staggered cases.