Don't store the last RK stage source
This stage source is only used implicitly and is never used in subsequent steps,
so there's no need to store it or to have a field for it in the stage source
list.

Also renamed stage sources to have a 'RK:' prefix, for clarity. And in the
default Poisson solver the initial guesses are initialized with the current
solution.