Removed boundary update at the start of solver's solve() function
There is no reason to update the solution's boundary values at the start of a
linear system solve. Boundary values are needed for an initial residual, but it
shouldn't matter too much if this is not very correct near the boundaries. And
also, both the pressure and velocity boundary values are up-to-date when calling
the solve function in briscolaColocated/Staggered solvers.

All validation cases seem the pass with this change.