Compute phiStar also in briscolaStaggeredTwoPhase
Instead of adding the reconstruction of the two-phase flux to the preliminary
staggered velocity solution and computing the colocated divergence of that, it
is better to first interpolate the preliminary staggered velocity solution to
colocated faces and adding the two-phase flux there. The benefits are that 1) no
boundary update is needed on U and 2) a call to stagReconstruct is avoided. The
drawback is that later on the two-phase flux needs to added again to the
staggered velocity field. However, the two-phase flux is a cached field anyway,
so it is not computed twice.

This change brings the implementation of the staggered solver closer to the
colocated solver. To improve similarity, the vcf field was renamed to vf.