Switched to boundary exchange mode
For higher number of processors, this communication mode, using MPI's
MPI_Neighbor_alltoallv function, seems to be more efficient. The
MPI_Neighbor_alltoallv uses a special communicator which is setup in the mesh
class and which uses weights for optimization. Also, boundary extension was
removed, which does not seem to have a negative effect on convergience.