Disabled memory zeroing in alloc()
This seems to give roughly a 10% efficiency increase. Note: currently the
staggered solvers fail with an RK scheme that is implicit for the convective
term. This is because the flux field is not restricted properly on staggered
meshes. It is probably better not to restrict face fields, but to caculate them
directly from cell-centered values that have their boundaries correct. But this
requires a multi-level parallel update which is still to be implemented.

To enable memory zeroing, the ZERO_INIT macro variable can be defined in
block.H. This is a useful option when debugging with the FOAM_SIGFPE environment
variable to catch divisions by zero other kind of sigFpe errors.