Better handling of deep and shallow fields
In field end set operations, return deep fields when possible. For functions
that require deep fields (i.e., implicit schemes and solvers), the rule is that
they should never change the original state of the field, meaning that a shallow
field should be made shallow again while a deep field should be left as is.