Improved staggered flux calculation
It is difficult to restrict staggered face fields, because coarse grid staggered
faces at boundaries fall outside the domain covered by the finer grid faces. So,
restricting the staggered phi field gives either wrong results near boundaries,
or requires expensive parallel boundary corrections. Instead, it is better to
compute the staggered face flux directly from the velocity on all levels. This
is now implemented. It is only needed if the convective term is treated
implicitly, so the restriction is only applied if rk.imA() is true.

With this, no staggered face restriction schemes are needed anymore. The
corresponding implementation is removed and replaced by a NotImplemented call,
to produce a warning when a staggered face field is restricted.